refactor: move api/papermag/decrypt into /pages/return/index
